Added frontend UI Integrations and

+++ backend/src/controller/order.ts
@@ -1,8 +1,19 @@
 import { Request, Response } from 'express';
 import prisma from '../config/db';
+import dotenv from 'dotenv';
+
+
+dotenv.config();
 
 // Create Order
 export const createOrder = async (req: Request, res: Response) => {
+    console.log('got here ooo')
+    const token = req.cookies.token;
+
+    console.log(token, 'token create orders')
+
+    if (!token) return res.status(401).json({ message: 'Unauthorized' });
+
     try {
         console.log(req.body, req?.user, 'got hahaaj')
         const { productName, productCategory, price, orderDate } = req.body;
@@ -12,7 +23,6 @@ export const createOrder = async (req: Request, res: Response) => {
         }
 
         const customerName = req.user.name;  // Access authenticated user's name
-        const customerId = req.user.userId;  // Access authenticated user's id
 
         const order = await prisma.order.create({
             data: {
@@ -32,13 +42,36 @@ export const createOrder = async (req: Request, res: Response) => {
 };
 
 // Get All Orders (Admin only)
-export const getAllOrders = async (_req: Request, res: Response) => {
-    const orders = await prisma.order.findMany();
-    return res.json(orders);
+export const getAllOrders = async (req: Request, res: Response) => {
+    // const PORT = process.env.PORT || 8000;
+    const token = req.cookies.token;
+
+    console.log(token, 'token get orders')
+
+    if (!token) return res.status(401).json({ message: 'Unauthorized' });
+    try {
+        // const user = jwt.verify(token, process.env.JWT_SECRET as string);
+        // user is now decoded { id, email, ... }
+
+        const orders = await prisma.order.findMany();
+        const totalOrders = await prisma.order.count();
+        return res.json({
+            data: orders,
+            total: totalOrders,
+        });
+    } catch (err) {
+        res.status(403).json({ message: 'Invalid token' });
+    }
+
 };
 
 // Update Order (Admin only)
 export const updateOrder = async (req: Request, res: Response) => {
+    const token = req.cookies.token;
+
+    console.log(token, 'token create orders')
+
+    if (!token) return res.status(401).json({ message: 'Unauthorized' });
     const { id } = req.params;
     const { productName, productCategory, price, orderDate } = req.body
     const order = await prisma.order.update({
@@ -56,6 +89,11 @@ export const updateOrder = async (req: Request, res: Response) => {
 
 // Delete Order (Admin only)
 export const deleteOrder = async (req: Request, res: Response) => {
+    const token = req.cookies.token;
+
+    console.log(token, 'token create orders')
+
+    if (!token) return res.status(401).json({ message: 'Unauthorized' });
     const { id } = req.params;
     await prisma.order.delete({ where: { id } });
